Update with DQL v1 on SDK 4.5.0-alpha.5
Refactored original demo app using the new DQL query language feature.

In the original implementation, the app centered around "Tasks". This commit also
renames the project, directory, and data model to use "ToDo"s. This is because the
`DittoStore.execute()` function in Swift is `async`, and calling an async function
in Swift uses the Swift `Task` object, the name of which collided with the original
data model name.